# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Brief": {
        "properties": {
          "audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Audience"
          },
          "business": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Business"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Goal"
          },
          "tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tone"
          }
        },
        "title": "Brief",
        "type": "object"
      },
...
      "CommandResponse": {
        "properties": {
          "brief": {
            "$ref": "#/components/schemas/Brief"
          },
          "clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Clarification"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "clarifications": {
            "items": {
              "type": "string"
            },
            "title": "Clarifications",
            "type": "array"
          },
          "commandId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Commandid"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Language"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageView"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "recognizedCommand": {
            "title": "Recognizedcommand",
            "type": "string"
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Sessionid"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionEntry"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "recognizedCommand",
          "message",
          "summary",
          "brief",
          "clarifications",
          "versions",
          "plan",
          "validation"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
...
      "EditChange": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          }
        },
        "required": [
          "action"
        ],
        "title": "EditChange",
        "type": "object"
      },
      "EditRequest": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "changes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EditChange"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changes"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "EditRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "recognizedCommand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recognizedcommand"
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenPage": {
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageView"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionEntry"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "page",
          "versions"
        ],
        "title": "OpenPage",
        "type": "object"
      },
...
      "PublicPage": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sections": {
            "items": {
              "discriminator": {
                "mapping": {
                  "about": "#/components/schemas/AboutSection",
                  "benefits": "#/components/schemas/BenefitsSection",
                  "contact": "#/components/schemas/ContactSection",
                  "cta": "#/components/schemas/CtaSection",
                  "faq": "#/components/schemas/FAQSection",
                  "footer": "#/components/schemas/FooterSection",
                  "gallery": "#/components/schemas/GallerySection",
                  "header": "#/components/schemas/HeaderSection",
                  "hero": "#/components/schemas/HeroSection",
                  "pricing": "#/components/schemas/PricingSection",
                  "process": "#/components/schemas/ProcessSection",
                  "promotion": "#/components/schemas/PromotionSection",
                  "services": "#/components/schemas/ServicesSection",
                  "stats": "#/components/schemas/StatsSection",
                  "team": "#/components/schemas/TeamSection",
                  "testimonials": "#/components/schemas/TestimonialsSection"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HeaderSection"
                },
                {
                  "$ref": "#/components/schemas/HeroSection"
                },
                {
                  "$ref": "#/components/schemas/AboutSection"
                },
                {
                  "$ref": "#/components/schemas/BenefitsSection"
                },
                {
                  "$ref": "#/components/schemas/ServicesSection"
                },
                {
                  "$ref": "#/components/schemas/ProcessSection"
                },
                {
                  "$ref": "#/components/schemas/StatsSection"
                },
                {
                  "$ref": "#/components/schemas/GallerySection"
                },
                {
                  "$ref": "#/components/schemas/TestimonialsSection"
                },
                {
                  "$ref": "#/components/schemas/PricingSection"
                },
                {
                  "$ref": "#/components/schemas/TeamSection"
                },
                {
                  "$ref": "#/components/schemas/PromotionSection"
                },
                {
                  "$ref": "#/components/schemas/FAQSection"
                },
                {
                  "$ref": "#/components/schemas/ContactSection"
                },
                {
                  "$ref": "#/components/schemas/CtaSection"
                },
                {
                  "$ref": "#/components/schemas/FooterSection"
                }
              ]
            },
            "title": "Sections",
            "type": "array"
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Theme"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "when": {
            "title": "When",
            "type": "string"
          }
        },
        "required": [
          "url",
          "version",
          "sections",
          "when"
        ],
        "title": "PublicPage",
        "type": "object"
      },
      "PublishRequest": {
        "properties": {
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "PublishRequest",
        "type": "object"
      },
      "PublishResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          },
          "readiness": {
            "$ref": "#/components/schemas/Readiness"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "when": {
            "title": "When",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "published",
          "version",
          "when",
          "readiness"
        ],
        "title": "PublishResponse",
        "type": "object"
      },
...
      "RenameRequest": {
        "properties": {
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "wanted": {
            "title": "Wanted",
            "type": "string"
          }
        },
        "required": [
          "page_id",
          "wanted"
        ],
        "title": "RenameRequest",
        "type": "object"
      },
      "RenameResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "published"
        ],
        "title": "RenameResponse",
        "type": "object"
      },
      "RestoreRequest": {
        "properties": {
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "page_id",
          "version"
        ],
        "title": "RestoreRequest",
        "type": "object"
      },
      "RestoreResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "page": {
            "$ref": "#/components/schemas/PageView"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionEntry"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "page",
          "versions",
          "message"
        ],
        "title": "RestoreResponse",
        "type": "object"
      },
...
      "UnpublishRequest": {
        "properties": {
          "page_id": {
            "title": "Page Id",
            "type": "string"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "UnpublishRequest",
        "type": "object"
      },
      "UnpublishResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "published"
        ],
        "title": "UnpublishResponse",
        "type": "object"
      },
...
      "ValidationStatus": {
        "properties": {
          "gates": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gates"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ValidationStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Magic Landing Pages",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/edit": {
      "post": {
        "operationId": "edit_edit_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Edit",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/pages/url": {
      "post": {
        "operationId": "rename_page_pages_url_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Page",
        "tags": [
          "pages"
        ]
      }
    },
    "/pages/{page_id}": {
      "get": {
        "operationId": "open_page_pages__page_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Page",
        "tags": [
          "pages"
        ]
      }
    },
    "/public": {
      "get": {
        "operationId": "public_page_public_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "title": "Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Page",
        "tags": [
          "pages"
        ]
      }
    },
    "/publish": {
      "post": {
        "operationId": "publish_publish_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Publish",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/unpublish": {
      "post": {
        "operationId": "unpublish_unpublish_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnpublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnpublishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unpublish",
        "tags": [
          "pages"
        ]
      }
    },
    "/versions/restore": {
      "post": {
        "operationId": "restore_versions_restore_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore",
        "tags": [
          "pages"
        ]
      }
    },
    "/versions/{version}": {
      "get": {
        "operationId": "version_snapshot_versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Version Snapshot",
        "tags": [
          "pages"
        ]
      }
    }
  }
}